implementation of feature 43
